Päivitetty: Tapahtuma(entity+dokumentaatio), tietokantakaavio + kuva, Tapahtuman delete- ja post-metodit, API dokumentaatio

+++ src/main/java/ohjelmistoprojekti1/a3004/web/RestTapahtumaController.java
@@ -1,6 +1,10 @@
 package ohjelmistoprojekti1.a3004.web;
 
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -23,15 +27,28 @@ public class RestTapahtumaController {
         return tapahtumaRepository.findAll();
     }
 
-    // haetaan tapahtumat, joiden alku_pvm on tulevaisuudessa
+    // haetaan tapahtumat, joiden alku on kuluvan vuorokauden jälkeen
     @GetMapping("/tapahtumat/tulevat")
     public Iterable<Tapahtuma> tulevatTapahtumat() {
-        return tapahtumaRepository.findByTuleva(true);
+        // haetaan vertailuajaksi kuluvan vuorokauden viimeinen hetki
+        LocalDateTime tanaan = LocalDateTime.now().with(LocalTime.MAX);
+        return tapahtumaRepository.findAllByAlkuAfter(tanaan);
     }
 
+    // @PostMapping("/tapahtumat")
+    // Tapahtuma uusiTapahtuma(@RequestBody Tapahtuma uusiTapahtuma) {
+    //     return tapahtumaRepository.save(uusiTapahtuma);
+    // }
+
     @PostMapping("/tapahtumat")
-    Tapahtuma uusiTapahtuma(@RequestBody Tapahtuma uusiTapahtuma) {
-        return tapahtumaRepository.save(uusiTapahtuma);
+    public ResponseEntity<Tapahtuma> uusiTapahtuma(@RequestBody Tapahtuma uusiTapahtuma) {
+        // tarkistetaan, onko tapahtumalle annettu nimi. Jos ei, palautetaan 400 - bad request
+        if (uusiTapahtuma.getTapahtuman_nimi() == null || uusiTapahtuma.getTapahtuman_nimi().isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        }
+        // jos tapahtumalle on annettu nimi, palautetaan 200 - ok
+        Tapahtuma tallennettuTapahtuma = tapahtumaRepository.save(uusiTapahtuma);
+        return ResponseEntity.ok(tallennettuTapahtuma);
     }
 
     @PutMapping("/tapahtumat/{id}")
@@ -40,9 +57,21 @@ public class RestTapahtumaController {
         return tapahtumaRepository.save(muokattuTapahtuma);
     }
 
-    @DeleteMapping("/tapahtumat/{id}")
-    public void poistaTapahtuma(@PathVariable("id") Long id) {
-        tapahtumaRepository.deleteById(id);
-    }
+    // @DeleteMapping("/tapahtumat/{id}")
+    // public void poistaTapahtuma(@PathVariable("id") Long id) {
+    //     tapahtumaRepository.deleteById(id);
+    // }
     
+    @DeleteMapping("/tapahtumat/{id}")
+    public ResponseEntity<Object> poistaTapahtuma(@PathVariable("id") Long id) {
+        // tarkistetaan, löytyykö tietokannasta tietuetta pyydetyllä id:llä
+        if (!tapahtumaRepository.existsById(id)) {
+            // jos ei löydy, palautetaan koodi 404 - not found 
+            return ResponseEntity.notFound().build();
+        }
+        // jos tietue löytyy, se poistetaan ja palautetaan koodi 200 - ok
+        tapahtumaRepository.deleteById(id);
+        return ResponseEntity.ok().build();
+    }
+
 }
